+++ changelog.md
@@ -1,4 +1,18 @@
-v0.0.2:
-- tilling
-- indexmap for windows map instead of hashmap (now order will be saved in 
-re-tilling)
+v0.0.3:
+- shortcuts handling
+- parsing shortcuts from config, e.g.:
+```toml
+[general]
+mainmod = "super"
+
+[shortcuts]
+"mod+t" = "xterm"
+```
+- new config var in `general` table: `sh`. String, determines prefed shell.
+E.g.:
+```toml
+[general]
+...
+sh = "zsh"
+```
+- small code cleanse (need more though?)